225 bringup
This is a much, much more functional version of UnrealRS2!

Uses Lesleys C client (Rev 225)
This makes the client side virtually complete.

Unreal can now start the client in-process, and on-thread.

The client currently only supports SDL2 presentation (separate window), but unreal has access to client data, and Unreal vanilla presentation is close!

After that, we can move on to Unreal processing the scene graph!